#22 Refactoring a movie interface.
