+++ src/main/java/org/aivinog1/omdb/api/movie/Coordinate.java
@@ -21,35 +21,30 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.aivinog1.omdb.api.movie.rating;
+package org.aivinog1.omdb.api.movie;
+
 /**
- * This is a base interface to interact with a
- * {@link org.aivinog1.omdb.api.movie.rating.Rating}.
+ * This is a basic interface to working with a movie identifier.
+ * For example: IMDB id.
  *
  * @author Alexey Vinogradov (vinogradov.a.i.93@gmail.com)
  * @version $Id$
- * @see org.aivinog1.omdb.api.movie.rating.Rating
+ * @see org.aivinog1.omdb.api.movie.Movie
  * @since 0.0.1
- * @todo #5:30m Let's implement this method.
- *  It should works with a numerous ratings.
+ * @todo #5:30m We need to implement this interface.
+ *  It should work with a movie coordinate.
  */
-public interface Ratings {
+public interface Coordinate {
     /**
-     * Iterating through Ratings.
+     * IMDB's movie identifier.
      *
-     * @return An available rating.
+     * @return A string representation of IDMB movie identifier.
      */
-    Iterable<Rating> iterate();
+    String identifier();
     /**
-     * A movie rating. E.G. PG-13.
+     * A movie website.
      *
-     * @return A string representation of a movie's rating.
+     * @return A string representation of a movie website.
      */
-    String rated();
-    /**
-     * Votes of a movie.
-     *
-     * @return A string representation of movie's votes.
-     */
-    String votes();
+    String website();
 }
