+++ src/main/java/org/aivinog1/omdb/api/movie/contributor/actor/package-info.java
@@ -21,35 +21,14 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.aivinog1.omdb.api.movie.rating;
 /**
- * This is a base interface to interact with a
- * {@link org.aivinog1.omdb.api.movie.rating.Rating}.
+ * This is a base package for an actor.
+ * In this package we can get an
+ * {@link org.aivinog1.omdb.api.movie.contributor.actor.Actor}
+ * from {@link org.aivinog1.omdb.api.movie.contributor.actor.Actors} variety.
  *
- * @author Alexey Vinogradov (vinogradov.a.i.93@gmail.com)
- * @version $Id$
- * @see org.aivinog1.omdb.api.movie.rating.Rating
+ * @author aivinog1
+ * @version 0.0.1
  * @since 0.0.1
- * @todo #5:30m Let's implement this method.
- *  It should works with a numerous ratings.
  */
-public interface Ratings {
-    /**
-     * Iterating through Ratings.
-     *
-     * @return An available rating.
-     */
-    Iterable<Rating> iterate();
-    /**
-     * A movie rating. E.G. PG-13.
-     *
-     * @return A string representation of a movie's rating.
-     */
-    String rated();
-    /**
-     * Votes of a movie.
-     *
-     * @return A string representation of movie's votes.
-     */
-    String votes();
-}
+package org.aivinog1.omdb.api.movie.contributor.actor;
